requirements.txt
snowflake-connector-python[secure-local-storage]
python-dotenv
sqlparse
